SD-1645: Recursive Decks
This uses the opaque component machinery to break the cycle in types
when trying to make a recursive components. Consequently, Decks can only
be peeked at the top-level, as everything is is now `Private`.

To allow child cards to introduce a recursive deck, we thread the
deck's component spec through as an argument.